chore(EnhancedAlert): convert to TS
